refactor(db) 6/13: extract db/ability.rs (closes #252)
Move 19 ability/talent/discipline/GSF populators (+ migrate_disciplines_from_dis_columns),
6 classifiers, the AbilityStats struct+impl, their CREATE TABLEs, and 24 tests into
db/ability.rs. Shared test helpers insert_obj/seed_combat_styles_for moved to testutil
(pub(crate)). 240 tests + clippy -D warnings + fmt green.
